+++ security/migrations/0003_delete_commenteventlog.py
@@ -1,0 +1,13 @@
+from django.db import migrations
+
+
+class Migration(migrations.Migration):
+    dependencies = [
+        ('security', '0002_commenteventlog'),
+    ]
+
+    operations = [
+        migrations.DeleteModel(
+            name='CommentEventLog',
+        ),
+    ]